local.conf: adjust IMAGE_OVERHEAD_FACTOR for developer builds
The image overhead factor is based on the used space in the rootfs
images and adds a given percentage on top so that free space
is later available on the device. However, for our currenet
eMMC partitions the size must be larger than 1 GiB, so
reduce the scaling factor a little bit.

Signed-off-by: Simon Stürz <simon.stuerz@chargebyte.com>
Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>